Multiple viewport widgets.
