+++ test-apps/appui-test-app/appui-test-providers/src/appui-test-providers.ts
@@ -12,6 +12,7 @@ export * from "./ui/providers/FloatingWidgetsUiItemsProvider";
 export * from "./ui/providers/InspectUiItemInfoToolProvider";
 export * from "./ui/providers/MessageUiItemsProvider";
 export * from "./ui/providers/PreviewFeaturesToggleProvider";
+export * from "./ui/providers/WidgetContentProvider";
 
 export * from "./ui/frontstages/ComponentExamples";
 export * from "./ui/frontstages/ContentLayout";
